requirements.txt
# Google Pub/Sub
google-cloud-pubsub==2.19.0

# Database dependencies (SQLAlchemy ORM)
sqlalchemy==2.0.25
pg8000==1.31.2
cloud-sql-python-connector==1.13.0

# Utility dependencies
python-dotenv==1.0.0
pydantic==2.5.3
-e .